requirements.txt
feedparser==6.0.11
requests==2.31.0
atproto==0.0.46
deepl==1.17.0
python-dotenv==1.0.1
